refactor: restructure project directories and simplify file names
