+++ src/app/app_init.cpp
@@ -1,5 +1,7 @@
 #include <fstream>
 
+#include <nfd.h>
+
 #include "logger/logger.h"
 
 #include "app/app.h"
@@ -180,5 +182,39 @@ namespace App {
 
         // 5. Free context itself
         delete ctx;
+    }
+
+    void openZmxFileInZemax() {
+        nfdchar_t* outPath = nullptr;
+        nfdresult_t result = NFD_OpenDialog("zmx", nullptr, &outPath);
+
+        if (result == NFD_OKAY) {
+        #ifdef DEBUG_LOG
+            logger.addLog(std::format("[APP] Selected file: {}", outPath));
+        #endif
+            // UTF-8 → UTF-16
+            int size = MultiByteToWideChar(CP_UTF8, 0, outPath, -1, nullptr, 0);
+            std::wstring widePath(size, L'\0');
+            MultiByteToWideChar(CP_UTF8, 0, outPath, -1, widePath.data(), size);
+
+            HINSTANCE ret = ShellExecuteW(nullptr, L"open", widePath.c_str(), nullptr, nullptr, SW_SHOW);
+            if (reinterpret_cast<intptr_t>(ret) <= 32) {
+                MessageBoxW(nullptr, L"Failed to open file. Is Zemax installed?", L"Error", MB_ICONERROR);
+                logger.addLog(std::format("[APP] ShellExecute failed to open file: {}. (Error code: {})", outPath, static_cast<int>(reinterpret_cast<intptr_t>(ret))));
+            } else {
+            #ifdef DEBUG_LOG
+                logger.addLog(std::format("[APP] Successfully sent file to ShellExecute: {}", outPath));
+            #endif
+            }
+
+            free(outPath);
+        } else if (result == NFD_CANCEL) {
+        #ifdef DEBUG_LOG
+            logger.addLog("[APP] File open dialog canceled by user");
+        #endif
+        } else {
+            const char* error = NFD_GetError();
+            logger.addLog(std::format("[APP] NFD Error: {}", error ? error : "Unknown error"));
+        }
     }
 }
