+++ src/dde/client_data.cpp
@@ -1,8 +1,8 @@
 #include <format>
 #include <dde.h>
 
-#include "dde_zemax_client.h"
-#include "dde_zemax_utils.h"
+#include "client.h"
+#include "utils.h"
 
 namespace ZemaxDDE {
     void ZemaxDDEClient::getLensName() {
