+++ src/dde/constants.h
