+++ src/dde/utils.h
