--- src/gui/components/gui_content.h
@@ -1,6 +1,0 @@
-#pragma once
-
-namespace gui {
-    inline constexpr float CONTENT_WIDTH_MIN    = 600.0f;
-    inline constexpr float CONTENT_HEIGHT_MIN   = 480.0f;
-}
