--- src/gui/components/gui_dde_status.h
@@ -1,23 +1,0 @@
-#pragma once
-
-#include "lib/imgui/imgui.h"
-
-namespace gui {
-    // Constants defining frame size
-    inline constexpr float DDE_STATUS_FRAME_WIDTH = 0.0f;       // Automatic width
-    inline constexpr float DDE_STATUS_FRAME_HEIGHT = 0.0f;      // Automatic height
-
-    // Status text color
-    inline constexpr ImVec4 DDE_STATUS_COLOR_CONNECTED = ImVec4(0.0f, 1.0f, 0.0f, 1.0f);            // Green
-    inline constexpr ImVec4 DDE_STATUS_COLOR_DISCONNECTED = ImVec4(1.0f, 0.0f, 0.0f, 1.0f);         // Red
-
-    // Button colors for the "Connect" state
-    inline constexpr ImVec4 DDE_BUTTON_CONNECT_COLOR_NORMAL = ImVec4(0.0f, 0.5f, 0.0f, 1.0f);       // Normal state
-    inline constexpr ImVec4 DDE_BUTTON_CONNECT_COLOR_HOVER = ImVec4(0.0f, 0.7f, 0.0f, 1.0f);        // Hovered state
-    inline constexpr ImVec4 DDE_BUTTON_CONNECT_COLOR_ACTIVE = ImVec4(0.0f, 0.3f, 0.0f, 1.0f);       // Active state
-
-    // Button colors for the "Disconnect" state
-    inline constexpr ImVec4 DDE_BUTTON_DISCONNECT_COLOR_NORMAL = ImVec4(0.5f, 0.0f, 0.0f, 1.0f);    // Normal state
-    inline constexpr ImVec4 DDE_BUTTON_DISCONNECT_COLOR_HOVER = ImVec4(0.7f, 0.0f, 0.0f, 1.0f);     // Hovered state
-    inline constexpr ImVec4 DDE_BUTTON_DISCONNECT_COLOR_ACTIVE = ImVec4(0.3f, 0.0f, 0.0f, 1.0f);    // Active state
-}
