--- src/gui/components/gui_debug_log.h
@@ -1,6 +1,0 @@
-#pragma once
-
-namespace gui {
-    inline constexpr float DEBUG_LOG_WIDTH_MIN  = 600.0f;
-    inline constexpr float DEBUG_LOG_HEIGHT_MIN = 200.0f;
-}
