--- src/gui/components/gui_navbar.h
@@ -1,7 +1,0 @@
-#pragma once
-
-/**
- * @file gui_navbar.h
- * @brief Top menu bar rendering (implemented in gui_navbar.cpp as GuiManager::renderNavbar()).
- *        This header exists as a module marker; no standalone declarations needed.
- */
