--- src/gui/components/gui_popups.h
@@ -1,6 +1,0 @@
-#pragma once
-
-namespace gui {
-    class GuiManager;  // Forward declaration
-    
-}
