--- src/gui/components/gui_sidebar.h
@@ -1,6 +1,0 @@
-#pragma once
-
-namespace gui {
-    inline constexpr float SIDEBAR_WIDTH_MIN    = 248.0f;
-    inline constexpr float SIDEBAR_HEIGHT_MIN   = 260.0f;
-}
