+++ src/gui/components/sidebar.cpp
@@ -1,6 +1,5 @@
-#include "gui/components/gui_sidebar.h"
-
 #include "gui/gui.h"
+#include "gui/constants.h"
 
 namespace gui {
     void GuiManager::renderSidebar() {
