--- src/gui/content_pages/gui_page_optical_system_info.h
@@ -1,8 +1,0 @@
-#pragma once
-
-/**
- * @file gui_page_optical_system_info.h
- * @brief Optical system information page rendering (implemented in gui_page_optical_system_info.cpp
- *        as GuiManager::renderPageOpticalSystemInfo()).
- *        This header exists as a module marker; no standalone declarations needed.
- */
