+++ src/gui/content_pages/page_optical_system_info.cpp
