+++ src/gui/content_pages/page_surface_sag_analysis.h
@@ -1,0 +1,11 @@
+#pragma once
+
+struct SurfaceSagAnalysisPageState {
+    int tolerancedSurfaceIndex = 0;
+    int nominalSurfaceIndex = 0;
+
+    int tolerancedSampling = 128;
+    double tolerancedAngle = 0.0;
+    int nominalSampling = 128;
+    double nominalAngle = 0.0;
+};
