+++ src/gui/init.cpp
