+++ src/gui/surface_analysis.cpp
