+++ src/gui/utils.cpp
